+++ ext/jtt/xtream-codec-ext-jt-808-server-spring-boot-starter/src/main/java/io/github/hylexus/xtream/codec/ext/jt808/dashboard/controller/BuiltinJt808DashboardSessionController.java
@@ -93,7 +93,7 @@ public class BuiltinJt808DashboardSessionController {
     Predicate<Jt808Session> createFilter(Jt808SessionQueryDto dto) {
         Predicate<Jt808Session> predicate = it -> true;
         if (dto.getTerminalId() != null) {
-            predicate = predicate.and(it -> it.terminalId().equals(dto.getTerminalId()));
+            predicate = predicate.and(it -> it.terminalId().toLowerCase().contains(dto.getTerminalId().toLowerCase()));
         }
         if (dto.getServerType() != null) {
             predicate = predicate.and(it -> it.role() == dto.getServerType());