+++ samples/jtt/jt-808-server-spring-boot-starter-debug/src/main/java/io/github/hylexus/xtream/debug/ext/jt808/controller/DemoJt808SessionController.java
@@ -16,69 +16,14 @@
 
 package io.github.hylexus.xtream.debug.ext.jt808.controller;
 
-import io.github.hylexus.xtream.codec.ext.jt808.spec.Jt808Session;
-import io.github.hylexus.xtream.codec.ext.jt808.spec.Jt808SessionManager;
-import io.github.hylexus.xtream.debug.ext.jt808.domain.dto.Jt808SessionQueryDto;
-import io.github.hylexus.xtream.debug.ext.jt808.domain.vo.Jt808SessionVo;
-import io.github.hylexus.xtream.debug.ext.jt808.domain.vo.PageableVo;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-
-import java.util.List;
-import java.util.function.Predicate;
 
 /**
  * @author hylexus
  */
 @RestController
-@RequestMapping("/api/v1")
+@RequestMapping("/demo-api/v1")
 public class DemoJt808SessionController {
-    private final Jt808SessionManager jt808SessionManager;
 
-    public DemoJt808SessionController(Jt808SessionManager jt808SessionManager) {
-        this.jt808SessionManager = jt808SessionManager;
-    }
-
-    @GetMapping("/instruction-session/list")
-    public PageableVo<Jt808SessionVo> list(Jt808SessionQueryDto dto) {
-        final Predicate<Jt808Session> filter = createFilter(dto);
-        final long total = this.jt808SessionManager.count(filter);
-        if (total <= 0) {
-            return PageableVo.empty();
-        }
-        final List<Jt808SessionVo> list = jt808SessionManager
-                .list(dto.getPage(), dto.getSize(), filter, this::convertToVo)
-                .toList();
-        return PageableVo.of(total, list);
-    }
-
-    Predicate<Jt808Session> createFilter(Jt808SessionQueryDto dto) {
-        Predicate<Jt808Session> predicate = it -> true;
-        if (dto.getTerminalId() != null) {
-            predicate = predicate.and(it -> it.terminalId().equals(dto.getTerminalId()));
-        }
-        if (dto.getServerType() != null) {
-            predicate = predicate.and(it -> it.role() == dto.getServerType());
-        }
-        if (dto.getProtocolVersion() != null) {
-            predicate = predicate.and(it -> it.protocolVersion() == dto.getProtocolVersion());
-        }
-        if (dto.getProtocolType() != null) {
-            predicate = predicate.and(it -> it.type() == dto.getProtocolType());
-        }
-        return predicate;
-    }
-
-    Jt808SessionVo convertToVo(Jt808Session it) {
-        return new Jt808SessionVo()
-                .setId(it.id())
-                .setTerminalId(it.terminalId())
-                .setServerType(it.role())
-                .setProtocolVersion(it.protocolVersion())
-                .setProtocolType(it.type())
-                .setCreationTime(it.creationTime())
-                .setLastCommunicateTime(it.lastCommunicateTime())
-                ;
-    }
 }