+++ CHANGELOG.md
@@ -2,7 +2,12 @@
 
 ### ⭐ New Features
 
-- 优化会话管理
+- 优化 `Jt808SessionManager`
+- 优化 `Jt808CommandSender`
+- 优化 `XtreamEventPublisher`
+- 内置 **actuator** 指标
+- 请求解码逻辑移动到 `Filter` 之前
+- 新增 `RequestDispatcherSchedulerFilter`
 
 ### ⚠️ Breaking Changes