+++ CHANGELOG.md
@@ -35,7 +35,7 @@
 新增两个子模块:
 
 - **xtream-codec-server-reactive**: 和具体协议格式无关的 **纯异步**、<span style="color:red;">非阻塞的</span> 的服务端
-- **xtream-codec-ext-jt-808-server-spring-boot-starter**: 基于 **xtream-codec-server-reactive** 实现的 **JT/T 808** 服务端扩展
+- **jt-808-server-spring-boot-starter-reactive**: 基于 **xtream-codec-server-reactive** 实现的 **JT/T 808** 服务端扩展
 
 ### ⚠️ Breaking Changes